requirements.txt
sounddevice
numpy
pyautogui
pillow
openai-whisper
torch
